renamed getServices to attachServices basic frontend work
