+++ app/firebase/functions/src/instances/app.ts
@@ -1,9 +1,9 @@
 import cors from 'cors';
 import express from 'express';
 
+import { attachServices } from '../middleware/attach.services';
 import { authenticate } from '../middleware/authenticate';
 import { errorHandling } from '../middleware/errorHandlingMiddleware';
-import { attachServices } from '../middleware/services';
 
 export const buildApp = (router?: express.Router): express.Application => {
   const app = express();
